requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
executing==2.2.0
fastapi==0.116.1
flatbuffers==25.2.10
fonttools==4.59.0
gast==0.6.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kagglehub==0.3.12
keras==3.10.0
keras-hub==0.21.1
keras-nlp==0.21.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.1
namex==0.1.0
nest-asyncio==1.6.0
numpy===None
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.2
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
split-folders==0.5.1
stack-data==0.6.3
starlette==0.47.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-text==2.16.1
termcolor==3.1.0
threadpoolctl==3.6.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
